functioning version finished
