+++ pyiron/sphinx/interactive.py
@@ -9,7 +9,7 @@ import subprocess
 import warnings
 import time
 from pyiron.sphinx.base import SphinxBase, Group
-from pyiron.atomistics.job.interactive import GenericInteractive
+from pyiron.atomistics.job.interactive import GenericInteractive, GenericInteractiveOutput
 from collections import OrderedDict as odict
 
 BOHR_TO_ANGSTROM = (
@@ -33,6 +33,7 @@ __date__ = "Sep 1, 2017"
 class SphinxInteractive(SphinxBase, GenericInteractive):
     def __init__(self, project, job_name):
         super(SphinxInteractive, self).__init__(project, job_name)
+        self.output = SphinxOutput(job=self)
         self._interactive_write_input_files = True
         self._interactive_library_read = None
         self._interactive_fetch_completed = True
@@ -389,6 +390,48 @@ class SphinxInteractive(SphinxBase, GenericInteractive):
             self.input.sphinx.main.extControl.bornOppenheimer = commands
         else:
             super(SphinxInteractive, self).load_main_group()
+
+
+class SphinxOutput(GenericInteractiveOutput):
+    def __init__(self, job):
+        super(SphinxOutput, self).__init__(job)
+
+    def check_band_occupancy(self, plot=True):
+        """
+            Check whether there are still empty bands available.
+
+            args:
+                plot (bool): plots occupancy of the last step
+
+            returns:
+                True if there are still empty bands
+        """
+        import matplotlib.pylab as plt
+        elec_dict = self._job['output/generic/dft']['n_valence']
+        if elec_dict is None:
+            raise AssertionError('Number of electrons not parsed')
+        n_elec = np.sum([elec_dict[k]
+                         for k in self._job.structure.get_chemical_symbols()])
+        n_elec = int(np.ceil(n_elec/2))
+        bands = self._job['output/generic/dft/bands_occ'][-1]
+        bands = bands.reshape(-1, bands.shape[-1])
+        max_occ = np.sum(bands>0, axis=-1).max()
+        n_bands = bands.shape[-1]
+        if plot:
+            xticks = np.arange(1, n_bands+1)
+            plt.xlabel('Electron number')
+            plt.ylabel('Occupancy')
+            plt.axvline(n_elec, label='#electrons: {}'.format(n_elec))
+            plt.axvline(max_occ, color='red',
+                label='Max occupancy: {}'.format(max_occ))
+            plt.axvline(n_bands, color='green',
+                label='Number of bands: {}'.format(n_bands))
+            plt.plot(xticks, bands.T, 'x', color='black')
+            plt.legend()
+        if max_occ < n_bands:
+            return True
+        else:
+            return False
 
 
 class SphinxInt2(SphinxInteractive):
